Justify or remove every hard-coded dtype in the alignment package
dev's dtype-conformance guard requires each torch.<dtype> literal outside
the exempt kernels to say why it deviates from the configured dtype. Of the
109 sites it flagged after the merge, five were unjustified and now use the
configured dtype: the two casts to double around the empirical sigma_A
ratio, the dense P1 transform's amplitudes, and the translation stage's
resolution mask and peak translations. The rest are annotated: index
tensors that index_add_ and gather require in int64, host-side 3x3 rotation
algebra in double, the rotation function's deliberate double accumulation
of oscillatory sums, exact clustering keys, and the anisotropy fit.

Full unit gate 2043 passed, 82 skipped (job 550854); alignment tests 192
(550852); panel 30/30 at both windows (550853).

+++ torchref/experimental/alignment/frf/api.py
@@ -371,8 +371,8 @@ class FastRotationFunction:
             # solvent deficit is simply what the ratio measures, per structure,
             # instead of two universal constants.
             eterm = empirical_sigma_a(
-                self._conv_obs.evaluate(smag_calc).to(torch.float64),
-                self._conv_calc.evaluate(smag_calc).to(torch.float64),
+                self._conv_obs.evaluate(smag_calc),
+                self._conv_calc.evaluate(smag_calc),
             ).to(F_calc.dtype)
         elif sigma_a_source == "luzzati":
             eterm = eterm_sigma_a(smag_calc, self.delta_vrms_A)
